a simulation created by open AI
This simulation was created by open AI with the command "create a simulation in html that simulates a helicopter taking off from a surface" 
I had to change the image to a .gif to make it seem more realistic. For the most part the AI did a good job producing the correct code.